refactor(docs): update helm instructions and values
As a pre-step towards 3.0 and enabling CSI drivers for jiva and
cstor, made the necessary documentation and values.yaml updates.

The premise is that openebs chart will continue its current templates
of ndm and local provisioner, while jiva and cstor will be deprecated
in favor of the corresponding csi charts.

Once 3.0 is released, the recommended way to install cstor, jiva and other
engines would be to enable via this umbrella chart. If there is a strong
use-case for using only one type of engine in a cluster - the user can opt to
install the sub-charts directly.

Trimmed down the configuration options in the README to the commonly used
options.

Also, added Shovan as a maintainer.

Signed-off-by: kmova <kiran.mova@mayadata.io>