chore(ci): rename 2.x to main in ci actions
Having two branches like main and 2.x pushing the images
doesn't work well with helm chart releaser when users
tend install with latest. A 2.12.5 can be shown as latst
if 2.12.5 chart is released post 3.0.0.

We will have the legacy components continue to maintain
in the main chart itself under the older version.

Signed-off-by: kmova <kiran.mova@mayadata.io>